remove some repetitions and increase clarity (#76)
I have read the CLA Document and I hereby sign the CLA

# Description

I reworked the first and second paragraphs. I removed the reference to
obscure "economic concepts" such as uniform clearing prices and
coincidence of wants. Now, we get faster to what they are with less
abstraction.

# Changes

dropped the last sentence of the first paragraph, rewrote the first
sentence of the second paragraph