Add technical specifications for hooks (#171)
# Description

We only have a "product feature" description for hooks. It's nicely
written but it lacks a few technical details.

This PR splits that page in two: high level feature description and
technical features". The technical features are expounded upon.

# Changes

- New technical page for hooks.
- Simplified hook feature page.
- Shared warning for critical knowledge.

## Related Tasks

PR #161 should be updated to link to the right page once this PR is
merged.