Update cow-amm.md (#331)
# Description

various edits


<!--- Describe your changes to provide context for reviewers, including
why it is needed -->

# Changes

<!-- List of detailed changes (how the change is accomplished) -->

- [x] the list of supported oracles had only Uniswap. But then in the
description there is also Balancer. I added Balancer to the initial list
- [x] edit the very last sentence (the problem with contributing only
one token to the AMM). It is not so much that the AMM will trade at an
unfavorable price: the price is always the one coming out of the
solvers' competition and should be correct. The problem is that it will
immediately sell a fraction of what it just received, which may not be
the intended behavior
- [x]  fixed some typos

<!--
## Related Issues

Fixes #
-->

---------

Co-authored-by: harisang <harisangelidakis@gmail.com>