chore: enforce toc when no section top-level content (#376)
# Description

When there is no content specified in a `README.mdx` / `README.md`
within a folder, this creates a "category" in Docusaurus, and, if there
is children content, this will just appear as navigation items under the
folder's name as a new Section (i.e. there is no TOC, and no specific
page for that folder / section when selected).

This causes issues, such as when wanting to refer users to "all order
types" (see https://docs.cow.fi/category/concepts and try to click
'Order Types'). It's a common request to be asked what order types are
supported, and not having a section that contains this in a snapshot
makes the documentation difficult to navigate / cumbersome.

# Changes

- [x] All sections that do not have a `README.mdx` / `README.md` (i.e.
no top level content), now have an auto-generated table of contents.