Update of CoW AMM solver tutorial (#410)
# Description

This is a quick update of the solver tutorial on integrating Balancer
CoW AMMs.

I am not really able to speak about the technical details. So it would
be good if someone from backend/smart contract could have a look.

The goal of the PR is to remove wrong information. Adding a good
tutorial can be done later.

# Changes

This PR mostly removes information which was outdated. In some places it
adds new links, contract names, functions.