Add action to remind reviewer about open PRs (#457)
# Description

We currently have a lot of open PRs in this repo which are either ready
to be reviewed or ready to be merged but not moving forward.

This PR adds a weekly job that posts a message with a link to open PRs
into slack to make it more visibile/easier for the relevant teams to
move it forward.

# Changes
- [x] add a weekly github action that posts open PRs into slack