Update Milkman repo links to actively maintained ones (#443)
# Description

In the Milkman page, we're currently linking to the original
repositories from the time of the grant.
Both repos are controlled by a third party and are not kept up-to-date
with the latest changes to the Milkman contract. In particular:
1. The bot/contracts aren't compatible with the bot we currently use in
prod.
2. The contracts contain a somewhat small security vulnerability and
should not be used as a base for new contracts.

# Changes

We update the links to point to the official version that we use in our
infrastructure.

# How to test

Try the new links, notice that they are managed by the DAO and have had
recent changes from known contributors.

# References

[Internal
discussion](https://cowservices.slack.com/archives/C0375NV72SC/p1736155757041969?thread_ts=1736150416.130649&cid=C0375NV72SC).