Update competition_rules.md (#498)
I removed the description of the batch auction and instead now discuss
the fair combinatorial auction.

Note the following:
- we didn't have a specific page where we described the batch auction,
and I kept this same structure. However, given that the fair
combinatorial auction is more complex, we may want to consider creating
a dedicated page for it
- I don't explain the specific algorithm we use to select winners, and
instead, I provide a link to the CIP. In my opinion, we could discuss
the winner selection algorithm if we introduce a specific page in the
docs dedicated to the fair combinatorial auction but not in the current
structure.
- I kept the discussion about EBBO since, technically, it is still a
rule

# Description

<!--- Describe your changes to provide context for reviewers, including
why it is needed -->

# Changes

<!-- List of detailed changes (how the change is accomplished) -->

- [ ] ...
- [ ] ...

<!--
## Related Issues

Fixes #
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Updated the competition rules to introduce "Uniform Directional
Clearing Prices (UDCP)" with new guidelines for order pricing and
exceptions for certain orders.
- Expanded the description of the auction process, including details on
scoring, ranking, bid selection, and reward mechanisms.
  - Clarified rules on deadlines and penalties for winning solvers.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>