doc(contracts/core): disclose compromised balancer contract vulnerability
We recently learned that due to Balancer Deployer wallet being compromised due to the profanity vanity address generator vulnerability.

As we are planning to continue deploying on these networks with the same deployer address, we wanted to disclose this as a known issue prior to deployment on these affected networks.