Autogenerate explorer links in address lists (#545)
# Description

Reviewing the addition or removal of new chains in the contract
addresses is always difficult to review. See for example [this
PR](https://github.com/cowprotocol/docs/issues/538): there are lines
with a thousand characters getting modified and it's very easy to miss
differences that could [lead to
issues](https://github.com/cowprotocol/docs/pull/543).

I tried two ways to work around this. One is better table formatting,
but the solution I found (idea from
[here](https://github.com/facebook/docusaurus/discussions/5155#discussioncomment-8036773))
requires relying on some relatively obscure package
[remark-grid-tables](https://github.com/zestedesavoir/zmarkdown/tree/master/packages/remark-grid-tables).
I didn't want to add an unknown dependency.

The other way was by relying on the fact that Docusaurus uses React.
This PR is my attempt to use React to make adding chains easier.

This PR is my proposal on how to tackle this issue. I've never done any
react before so the code is likely low quality and I don't fully
understand what I'm doing. Please review carefully and question every
choice!

Ideas for the future: it would be great to add the chain icon next to
the name, so it's easier to find the chain you need.

# Changes

- Added new function to build links to explorer for a single contract.
It has some minimal customization via option parameters.
- Used new function in every place I could. (There are still a lot of
"raw" Etherscan link around in .md files. I think that's ok, if we feel
the need we can migrate them to use this function.)

# Known issues

Lens explorer doesn't follow the same anchor format used by Etherscan,
unlike all other explorers.
This means that things like
`https://explorer.lens.xyz/address/0x9008D19f58AAbD9eD0D60971565AA8510560ab41#code`
doesn't link to the code, the anchor should be `#contract`.
I could add some code to handle this edge case, but for now I'd leave it
as it is since it's just Lens. The link still works, it's just that it
doesn't point to the code.

# How to test

Visually compare the preview of this PR with the current version. Click
on a few links to confirm they point to the expected URL.

List:
- Periphery:
[current](https://docs.cow.fi/cow-protocol/reference/contracts/periphery),
[new](https://docs-git-autogenerate-explorer-links-cowswap.vercel.app/cow-protocol/reference/contracts/periphery).
- Core:
[current](https://docs.cow.fi/cow-protocol/reference/contracts/core),
[new](https://docs-git-autogenerate-explorer-links-cowswap.vercel.app/cow-protocol/reference/contracts/core).
- Signing schemes:
[current](https://docs.cow.fi/cow-protocol/reference/core/signing-schemes),
[new](https://docs-git-autogenerate-explorer-links-cowswap.vercel.app/cow-protocol/reference/core/signing-schemes).